apps/api/pyproject.toml
[project]
name = "knowledge-copilot-api"
version = "0.1.0"
description = "Enterprise Knowledge Copilot Backend"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",
    # Redis / Workers
    "redis>=5.0.0",
    "arq>=0.26.0",
    # Vector DB
    "qdrant-client>=1.12.0",
    # AI Providers
    "openai>=1.50.0",
    "anthropic>=0.39.0",
    "tiktoken>=0.7.0",
    # Auth
    "msal>=1.30.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    # Document processing
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    # Azure
    "azure-storage-blob>=12.20.0",
    "azure-identity>=1.17.0",
    # Microsoft Graph (SharePoint)
    "msgraph-sdk>=1.5.0",
    # Utils
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.0.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "httpx>=0.27.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
